+++ Notes/changes.md
@@ -3,3 +3,10 @@
 ## Niko - 22.11
 + started the preprocessing scripts
 + trained the model to find the plates from the images
+
+## Niko - 23.11
++ Working preprocessing
+    + can take an image, and return a cropped plate to a specified save file
+    + see src/preprocessing/preprocessing.py and tests/test_detection.py
++ wrote testfiles
++ fixed some